requirements.txt
aiohttp @ file:///C:/b/abs_c4zmy2l696/croot/aiohttp_1670009573673/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
appdirs==1.4.4
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout @ file:///C:/b/abs_43ozhz2a8g/croots/recipe/async-timeout_1664876362767/work
attrs==23.1.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.6.2
Brotli==1.0.9
brotlipy==0.7.0
certifi==2023.7.22
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cli2==2.6.2
click==8.1.6
cmake==3.27.0
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
cryptography @ file:///C:/b/abs_13590mi9q9/croot/cryptography_1689373706078/work
ctransformers==0.2.16
dataclasses-json @ file:///C:/b/abs_edxyjf9i7x/croot/dataclasses-json_1682969367890/work
debugpy @ file:///D:/bld/debugpy_1680755654494/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1675116244235/work
docstring-parser==0.7.1
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
faiss==1.7.4
filelock @ file:///C:/b/abs_c7yrhs9uz2/croot/filelock_1672387617533/work
Flask==2.3.2
frozenlist @ file:///C:/b/abs_2bb5uzghsi/croot/frozenlist_1670004511812/work
fsspec @ file:///C:/b/abs_f9arcwflp9/croot/fsspec_1682526006955/work
greenlet @ file:///C:/b/abs_47lk_w2ajq/croot/greenlet_1670513248400/work
huggingface-hub @ file:///C:/b/abs_46tn6fu5ho/croot/huggingface_hub_1686693723427/work
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
ipykernel @ file:///D:/bld/ipykernel_1690311464685/work
ipython @ file:///C:/b/abs_d1yx5tjhli/croot/ipython_1680701887259/work
itsdangerous==2.1.2
jedi @ file:///C:/ci/jedi_1644315428305/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib @ file:///C:/b/abs_1anqjntpan/croot/joblib_1685113317150/work
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client @ file:///C:/b/abs_059idvdagk/croot/jupyter_client_1680171872444/work
jupyter_core @ file:///C:/b/abs_9d0ttho3bs/croot/jupyter_core_1679906581955/work
langchain @ file:///home/conda/feedstock_root/build_artifacts/langchain_1689963353703/work
langsmith @ file:///home/conda/feedstock_root/build_artifacts/langsmith_1690588732965/work
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
marshmallow @ file:///C:/b/abs_2a28q2ehfk/croot/marshmallow_1682620739649/work
marshmallow-enum @ file:///C:/ci_310/marshmallow-enum_1642674085091/work
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934094726/work
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath==1.2.1
multidict @ file:///C:/b/abs_6cx_8w3cv2/croot/multidict_1665674238352/work
mutagen==1.46.0
mypy-extensions==0.4.3
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
networkx @ file:///C:/b/abs_b935xy_9g6/croot/networkx_1678964342510/work
nltk @ file:///C:/b/abs_a638z6l1z0/croot/nltk_1688114186909/work
numexpr @ file:///C:/b/abs_afm0oewmmt/croot/numexpr_1683221839116/work
numpy @ file:///C:/b/abs_5akk51tu0f/croot/numpy_and_numpy_base_1687466253743/work
openai==0.27.8
openapi-schema-pydantic @ file:///C:/b/abs_096qwt5_27/croot/openapi-schema-pydantic_1682536037030/work
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py-cpuinfo==9.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.18.0
pydantic @ file:///C:/b/abs_4bum6ajkso/croot/pydantic_1686125732638/work
pydub @ file:///home/conda/feedstock_root/build_artifacts/pydub_1615612442567/work
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pypdf @ file:///home/conda/feedstock_root/build_artifacts/pypdf-split_1690813086583/work/dist
pyproject-toml==0.0.10
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pywin32==305.1
PyYAML @ file:///C:/b/abs_d0g7dqt2xw/croot/pyyaml_1670514768165/work
pyzmq @ file:///C:/b/abs_655zk4a3s8/croot/pyzmq_1686601465034/work
referencing==0.30.0
regex @ file:///C:/ci/regex_1658258299320/work
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
rpds-py==0.9.2
scikit-build @ file:///home/conda/feedstock_root/build_artifacts/scikit-build_1685579843875/work
scikit-learn @ file:///C:/b/abs_38k7ridbgr/croot/scikit-learn_1684954723009/work
scipy==1.10.1
sentence-transformers @ file:///home/conda/feedstock_root/build_artifacts/sentence-transformers_1656297135090/work
sentencepiece @ file:///C:/b/abs_f7cttiucvr/croot/sentencepiece_1684525347071/work/python
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy @ file:///C:/Windows/Temp/abs_f8661157-660b-49bb-a790-69ab9f3b8f7c8a8s2psb/croots/recipe/sqlalchemy_1657867864564/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1684180539862/work
tenacity @ file:///C:/b/abs_ddkoa9nju6/croot/tenacity_1682972298929/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tokenizers @ file:///C:/b/abs_e34f9vyhry/croot/tokenizers_1687191954304/work
toml==0.10.2
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado @ file:///C:/ci/tornado_1662476985533/work
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
traitlets @ file:///C:/b/abs_e5m_xjjl94/croot/traitlets_1671143896266/work
transformers @ file:///C:/b/abs_8byf5_j714/croot/transformers_1667919454001/work
typing-inspect @ file:///tmp/build/80754af9/typing_inspect_1635544747198/work
typing_extensions @ file:///C:/b/abs_213vg2cd1l/croot/typing_extensions_1690297804941/work
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
websockets==11.0.3
Werkzeug==2.3.6
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
yarl @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e5nlunspj6/croots/recipe/yarl_1661437086516/work
youtube-dl @ file:///home/conda/feedstock_root/build_artifacts/youtube-dl_1640026230460/work
yt-dlp==2023.7.6
